# target/generated-sources/openapi/src/main/java/com/ahmetgogebakan/booklist/swaggergen/model/CheckDuplicationResponse.java
package com.ahmetgogebakan.booklist.swaggergen.model;

import java.util.Objects;
import com.ahmetgogebakan.booklist.swaggergen.model.CheckDuplicationResponseAllOf;
import com.ahmetgogebakan.booklist.swaggergen.model.Response;
import com.ahmetgogebakan.booklist.swaggergen.model.Result;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
    import java.io.Serializable;
    import javax.validation.Valid;
    import javax.validation.constraints.*;

            import java.util.Objects;
import io.swagger.v3.oas.annotations.media.Schema;

@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2020-12-27T02:04:12.306597+03:00[Europe/Istanbul]")

public class CheckDuplicationResponse implements Serializable {
        @Schema(example = "", description = "", pattern = "", enumAsRef = true)
        @JsonProperty("result")
    private Result result;

        @Schema(example = "Your transaction has been completed", description = "Result message", pattern = "", enumAsRef = true)
        @JsonProperty("message")
    private String message;

        @Schema(example = "2600000015131111", description = "Unique identifier of the user session", pattern = "", enumAsRef = true)
        @JsonProperty("sessionId")
    private String sessionId;

        @Schema(example = "true", description = "check for duplication", pattern = "", enumAsRef = true)
        @JsonProperty("isDuplicate")
    private Boolean isDuplicate;


    @ApiModelProperty(required = true)
    @NotNull(message="CheckDuplicationResponse.result.notNull")

public Result getResult() {
return result;
}

public void setResult(Result result) {
this.result = result;
}



public String getMessage() {
return message;
}

public void setMessage(String message) {
this.message = message;
}



public String getSessionId() {
return sessionId;
}

public void setSessionId(String sessionId) {
this.sessionId = sessionId;
}



public Boolean getIsDuplicate() {
return isDuplicate;
}

public void setIsDuplicate(Boolean isDuplicate) {
this.isDuplicate = isDuplicate;
}

@Override
public String toString() {
return "CheckDuplicationResponse{" +
", result=" + result +
", message=" + message +
", sessionId=" + sessionId +
", isDuplicate=" + isDuplicate +
"}";
}
}